change data structure and page directory
