--- hemle_groupwork/src/components/About/OurFriends/Pc.js
@@ -1,6 +1,0 @@
-const pc = ({tab}) => {
-    console.log( tab )
-    return tab.map(img => <img src={img} alt=''/>)
-}
-
-export default pc;
