--- hemle_groupwork/src/components/About/OurFriends/style.css
@@ -1,36 +1,0 @@
-#about_page #OurFriends {
-    flex-direction: column;
-    padding-bottom: 3em;
-}
-
-#about_page #OurFriends .ourfriend-content {
-    flex-direction: column;
-}
-
-
-
-#about_page #OurFriends .ourfriend-images {
-    height: 5em;
-    overflow: hidden;
-}
-
-#about_page #OurFriends .ourfriend-images .imgs {
-    display: flex;
-    align-items: center;
-    flex-wrap: nowrap;
-
-    height: 120%;
-    overflow-x: auto;
-    overflow-y: hidden;
-}
-
-#about_page #OurFriends .ourfriend-images img {
-    margin: 2em 5em;
-}
-
-
-@media all and (max-width: 454px) {
-    #about_page #OurFriends .ourfriend-article p {
-        line-height: 1.75em;
-    }
-}
