--- hemle_groupwork/src/components/About/OurTeams/style.css
@@ -1,83 +1,0 @@
-#about_page #OurTeams {
-    flex-direction: column;
-}
-
-#about_page #OurTeams .ourTeam-content {
-    flex-direction: column;
-}
-
-#about_page #OurTeams .ourTeam-images {
-    position: relative;
-    justify-content: center;
-    max-height: 29em;
-}
-
-#about_page #OurTeams .ourTeam-article p {
-    max-width: 35em;
-    line-height: 2em;
-}
-
-
-
-#about_page #OurTeams .slide {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    height: 347px;
-    margin: 1em 0;
-    position: absolute;
-}
-#about_page #OurTeams .slidediv {
-    display: flex;
-    justify-content: center;
-    align-items: center;
-
-    background: linear-gradient(to right, red, yellow, green);
-    padding: 1px;
-    border-radius: 100%;
-}
-#about_page #OurTeams .slide svg {
-    width: 2em;
-    height: 2em;
-    color: white;
-    
-    background: #292A2A;
-    border-radius: 100%;
-}
-#about_page #OurTeams .slide#left { left: -5%; }
-#about_page #OurTeams .slide#right { right: -5%; }
-
-
-#about_page #OurTeams .ourTeam-images .imgs {
-    display: flex;
-    justify-content: space-around;
-}
-
-
-
-#about_page #OurTeams .imgs .aside {
-    display: flex;
-    flex-direction: column;
-
-    width: 245px;
-    margin: 1em 1.5em;
-}
-
-#about_page #OurTeams .aside img {
-    width: 100%;
-    height: 347px;
-    border-radius: 10px;
-}
-
-#about_page #OurTeams .aside h2 {
-    margin: 20px 0;
-    font-size: 15px;
-    font-weight: bold;
-    color: black;
-}
-@media all and (max-width: 454px) {
-    #about_page #OurTeams .ourTeam-article p {
-        line-height: 1.75em;
-    }
-}
