--- hemle_groupwork/src/components/Cartes/data.jsx
@@ -1,66 +1,0 @@
-import img from "../../assets/tools/player.png"
-
-const theCards = [
-    {
-        position: "gardiens",
-        hg: "44",
-        players: [
-            {
-                img: img,
-                number: 1,
-                shortName: "Pierre Onana",
-                job: "Defenseur central",
-                fullName: "Arnaud Boris Mbala Bengono"
-            }, {
-                img: img,
-                number: 2,
-                shortName: "Pierre Onana",
-                job: "Defenseur central",
-                fullName: "Arnaud Boris Mbala Bengono"
-            }, {
-                img: img,
-                number: 999,
-                shortName: "Pierre Onana",
-                job: "Defenseur central",
-                fullName: "Arnaud Boris Mbala Bengono"
-            }, {
-                img: img,
-                number: 999,
-                shortName: "Pierre Onana",
-                job: "Defenseur central",
-                fullName: "Arnaud Boris Mbala Bengono"
-            }, {
-                img: img,
-                number: 999,
-                shortName: "Pierre Onana",
-                job: "Defenseur central",
-                fullName: "Arnaud Boris Mbala Bengono"
-            }, {
-                img: img,
-                number: 999,
-                shortName: "Pierre Onana",
-                job: "Defenseur central",
-                fullName: "Arnaud Boris Mbala Bengono"
-            }
-        ]
-    }, {
-        position: "attaquant",
-        players: [
-            {
-                img: img,
-                number: 999,
-                shortName: "Pierre Onana",
-                job: "Defenseur central",
-                fullName: "Arnaud Boris Mbala Bengono"
-            }, {
-                img: img,
-                number: 999,
-                shortName: "Boris Mbala",
-                job: "Boris Mbala",
-                fullName: "Arnaud Boris Mbala Bengono",
-            }
-        ]
-    }
-]
-
-export default theCards
