--- hemle_groupwork/src/components/Cartes/index.js
@@ -1,7 +1,0 @@
-import datas from "./data"
-import Section from "./Section"
-import "./style.css"
-
-const page = () => datas.map(tmp => <Section section={tmp}/> )
-
-export default page
