--- hemle_groupwork/src/components/Cartes/style.css
@@ -1,3 +1,0 @@
-body {
-    background: #3C3C3C;
-}
