--- hemle_groupwork/src/components/Infos/Description/index.js
@@ -1,14 +1,0 @@
-import "./style.css"
-
-const Section = ({desc}) =>
-    <section id="playerDescription">
-        
-        <h2> {desc.title} </h2>
-        
-        <p id="descriptionBody">
-            {desc.content}
-        </p>
-
-    </section>
-
-export default Section
