+++ hemle_groupwork/src/components/about_OurFriends/Pc.js
@@ -1,0 +1,8 @@
+import { SwiperSlide } from "swiper/react";
+
+const pc = ({tab}) =>
+    tab.map( (img, index) => 
+        <SwiperSlide key={index}> <img src={img} alt=''/> </SwiperSlide>
+    )
+
+export default pc;
