+++ hemle_groupwork/src/components/player_profil_bibliographie/index.js
@@ -1,0 +1,10 @@
+import "./style.css"
+
+const Section = (params) =>
+
+    <section className="playerBiographie">
+        <h2> {params.title} </h2>
+        <p> {params.description} </p>
+    </section>
+
+export default Section
