+++ hemle_groupwork/src/components/player_profil_presentation/style.css
@@ -1,0 +1,129 @@
+.player-presentation {
+    display: flex;
+    justify-content: center;
+    align-items: flex-end;
+    align-self: center;
+    flex-wrap: wrap;
+}
+
+
+
+
+
+.player-presentation-imageBox {
+    display: flex;
+    justify-content: center;
+
+    width: 319px;
+    height: 311px;
+    overflow: hidden;
+    position: relative;
+    box-sizing: content-box;
+    margin-right: 2em;
+}
+
+.player-presentation-imageBox img {
+    margin-top: 3em;
+}
+
+.player-presentation-imageGradient {
+    position: absolute;
+    inset: 0;
+    background: linear-gradient(#0000 0%, rgba(0, 0, 0, 0.76) 76.56%, #1F1414 100%);
+}
+
+.player-presentation .player-number {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+
+    background: url("../../assets/tools/circle.png");
+    background-size: 100% 100%;
+    width: 55px; height: 55px;
+    border-radius: 50%;
+    margin: 28px 20px;
+}
+
+
+
+
+
+
+
+.player-presentation-infoBox {
+    display: flex;
+    flex-direction: column;
+
+    flex: 1;
+    width: 319px;
+    max-width: 372px;
+    background: #292A2A;
+    padding: 0 1em;
+    color: white;
+}
+
+.player-presentation-infoTitle {
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    flex-wrap: wrap;
+
+    border-bottom: 1px solid white;
+    margin-bottom: 1em;
+    padding: 0 1em;
+}
+
+.player-presentation-infoTitle img {
+    width: 3em;
+    height: 3em;
+    border-radius: 50%;
+    margin-right: 10px;
+}
+
+.player-presentation-infoTitle h2 {
+    font-size: 18px;
+    padding: 25px 0;
+    text-align: center;
+}
+
+
+.player-presentation-infoList .label {
+    font-size: 14px;
+    font-weight: 600;
+    line-height: 45px;
+    text-transform: capitalize;
+}
+
+.player-presentation-infoList .value {
+    font-size: 15px;
+}
+
+
+
+
+
+
+.player-presentation:hover .player-presentation-imageBox img {
+    scale: 1.15;
+    transition: 1s;
+}
+.player-presentation:hover .player-presentation-infoTitle h2 {
+    color: yellow;
+    transition: 1s;
+}
+
+
+@media all and (max-width: 890px) {
+    .player-presentation-imageBox {
+        width: 100%;
+        margin-right: 0;
+    }
+    .player-presentation-imageBox .player-number {
+        margin: 12px 14px;
+        width: 40px;
+        height: 38px;
+    }
+    .player-presentation-infoBox {
+        max-width: 100%;
+    }
+}
