+++ hemle_groupwork/src/data/playersList.jsx
@@ -1,0 +1,31 @@
+import player_image_1 from "../assets/tools/player.png"
+
+const players = [
+    {
+        number: 1,
+        picture: player_image_1,
+        
+        fullname: "Arnaud Boris Mbala Bengono",
+        shortname: "Pierre Onana",
+        
+        position: "gardiens",
+        job: "gardien de but",
+        place: "Milieu de terrain droit",
+        club: 0,
+
+        birthday: "",
+        from: "Camerounaise",
+        kilo:'74',
+        tail:'179',
+
+        description: "Lorem Ipsum is simply dummy text of the printing and typesetting industry. " +
+            "Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when " +
+            "an unknown printer took a galley of type and scrambled it to make a type specimen " +
+            "book. It has survived not only five centuries, but also the leap into electronic " +
+            "typesetting, remaining essentially unchanged. It was popularised in the 1960s with " + 
+            "the release of Letraset sheets containing Lorem Ipsum passages, and more recently " +
+            "with desktop publishing software like Aldus PageMaker including versions"
+    }, 
+]
+
+export default players
