+++ hemle_groupwork/src/pages/player_list/style.css
@@ -1,0 +1,7 @@
+body {
+    background: #3C3C3C;
+}
+
+#root1 {
+    display: block;
+}
